Use `qlim` to defer messages from `TrackSelected`
We need to defer because we cannot update the `Id` of a `live.observer` in
response to a notification.

We could use `defer` or `deferlow`, but `qlim` has the additional benefit that
if the selection status toggles very rapidly, we'll only use the latest
version. Probably not really an issue, but may as well? Some nice info about
threading in M4L at

https://cycling74.com/tutorials/advanced-max-learning-about-threading