Guard against loop in preview
However, this is not the cause of the crash. Something in Frontend is causing
Live/Max to hang, but it's a nightmare to isolate, partly because it seems to
be affected also by some leftover state in the Max interpreter? Suddenly a
patch started workign again (one that I was minimizing to unearth the cause of
the hang), and then kept working, but when I then loaded the Frontend again --
hang; and then the _exact same patcher that was working before_ now started to
crash again. Infuriating.

Deleting

```
~/Library/Application Support/Cycling '74/Max 8
```

seems to affect it?