Process key-switching in the driver
We don't want to do this in the voices, because _if_ they would be run
concurrently, we might get into similar problems as we had before: a key-switch
command might be executed by one voice, and then the actual note might be
executed by another, while the first one hasn't completed yet -- and we'd have
a race condition again. One nice accidental benefit of this new setup is that
the structure becomes easier.