Add generic tonal backend
Might want to tweak this a bit more still, but this is suitable for playing any
handpan VST that offers a simple single chromatic tonal interface.

TODO: The velocity curves I'm seeing seem a bit odd; even at a hard tap on the
key, it's still pretty low. Should take a closer look.